requirements_colab.txt
pytorch-lightning
hydra-core
editdistance
pandas
omegaconf
